+++ lib-ltc/include/ltcparams.h
@@ -220,7 +220,7 @@ private:
     	static constexpr auto RTPMIDI = (1U << 7);
     };
 
-	void HandleDisabledOutput(const char *pLine, const char *pKeyword, unsigned nMaskDisabledOutputs);
+	void HandleDisabledOutput(const char *pLine, const char *pKeyword, uint8_t nMaskDisabledOutputs);
 	void SetBool(const uint8_t nValue, uint8_t& nProperty, const uint32_t nMask);
 	void SetValue(const bool bEvaluate, const uint8_t nValue, uint8_t& nProperty, const uint32_t nMask);
